Take ownership of environment files
While Rails really wants to control the coding style and contents of these
files, in particular during `rails app:upgrade`, these files are our domain.

We'll be adding changes to them, we'll be modifying them to fit our application,
so we want them to reflect our application, our coding styles.