Verify that we can eager load code
We've unfortunately deployed a file to production on a project that
couldn't be eager loaded, which broke the entire application, even
though nothing in development and CI warned us.

This is a step towards catching problems like the below, which would
otherwise only occur on boot in production mode:

   Hold on, I am eager loading the application.
   rake aborted!
   NameError: uninitialized constant Form::InputGroup::Component

      class DateComponent < Form::InputGroup::Component
                                             ^^^^^^^^^^^
   Did you mean?  Complex
   ~/frontlobby/app/components/tailwind_ui/application_ui/form/input_group/date_component.rb:5:in
   `<module:InputGroup>'

This also gives us a fine opportunity to check everything for
compatiblity with Zeitwerk.