Use orbs to install Node
Circle CIs node orb uses .nvmrc to determine the version of Node to
install, if one hasn't explicitly been specified using `node-version`.

However, we use .tool-versions to specify the version of Node to use, so
we have to manually extract the version from that file and write it to
.nvmrc for the orb to use.

See https://circleci.com/developer/orbs/orb/circleci/node for details.